refactor: 범용 Exception 커스텀으로 변환
