+++ src/main/java/shopping/domain/user/Password.java
@@ -1,5 +1,7 @@
 package shopping.domain.user;
 
+import shopping.exception.InvalidRequestException;
+
 import javax.persistence.Embeddable;
 
 @Embeddable
@@ -17,7 +19,7 @@ public class Password {
 
     private void validate(String password) {
         if (password == null) {
-            throw new IllegalArgumentException("비밀번호는 null일 수 없습니다.");
+            throw new InvalidRequestException("비밀번호는 null일 수 없습니다.");
         }
     }
